// car-parking-fe/src/app/model/i-car-in-out.ts
export interface ICarInOut {
  carId?: number;
  carPlateNumber?: string;
  carName?: string;
  carBrand?: string;
  carTypeName?: string;
  customerName?: string;
  customerPhoneNumber?: string;
  locationName?: string;
  floorName?: string;
  ticketEffectiveDate?: string;
  ticketExpiryDate?: string;
  carInOutId?: string;
  timeIn?: string;
  timeOut?: string;
  urlCarInImage?: string;
  urlCarOutImage?: string;
  customerIdCard?: string;
}